+++ src/App.tsx
@@ -1,8 +1,6 @@
 import { Route, Routes } from "react-router-dom";
 
-import AddItemForm from "./pages/AddItemForm";
 import Menu from "./pages/Menu";
-import Sales from "./pages/Sales";
 
 import "./App.css";
 import Items from "./pages/Items";
@@ -12,8 +10,6 @@ function App() {
 		<Routes>
 			<Route path="/" element={<Items />} />
 			<Route path="/menu" element={<Menu />} />
-			<Route path="/add-item" element={<AddItemForm />} />
-			<Route path="/sales" element={<Sales />} />
 		</Routes>
 	);
 }
