+++ src/hooks/useFirebaseRef.ts
@@ -1,6 +1,6 @@
 import React, { SyntheticEvent, forwardRef, useLayoutEffect, useState } from "react";
 import { DataSnapshot, onValue, push, ref, set } from "firebase/database";
-import database from "../../utils/firebase";
+import database from "../utils/firebase";
 
 const useFirebaseRef = (path: string) => {
     const [categories, setCategories] = useState<string[]>([]);
@@ -10,7 +10,7 @@ const useFirebaseRef = (path: string) => {
 			const data = snapshot.val();
 			setCategories(Object.values(data));
 		});
-	}, []);
+	});
 
     return categories;
 }
