--- src/pages/AddItemForm/index.tsx
@@ -1,3 +1,0 @@
-import AddItemForm from "./AddItemForm";
-
-export default AddItemForm;
