+++ src/pages/ItemForm/index.tsx
@@ -1,0 +1,3 @@
+import ItemForm from "./ItemForm";
+
+export default ItemForm;
