--- src/pages/Sales/Sales.tsx
@@ -1,5 +1,0 @@
-const Sales = () => {
-    return <h1>Sales</h1>
-}
-
-export default Sales;
