--- src/pages/Sales/index.tsx
@@ -1,3 +1,0 @@
-import Sales from "./Sales";
-
-export default Sales;
